pyproject.toml
[project]
name = "remote-read-sql"
version = "1.0.2"
description = "Securely read sql into a pandas dataframe from a remote mysql DB"
readme = "README.rst"
keywords = ["pandas", "ssh", "mysql", "remote", "dataframe"]
authors = [
    { name = "Erik van Widenfelt", email = "ew2789@gmail.com" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.13",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
requires-python = ">=3.12"
dependencies = [
    "mysqlclient>=2.2.7",
    "pandas>=2.3.3",
    "paramiko==2.9.3",
    "python-dotenv>=1.1.1",
    "sphinx>=8.2.3",
    "sqlalchemy>=2.0.44",
    "sqlglot>=27.28.1",
    "sshtunnel>=0.4.0",
]

[build-system]
requires = ["uv_build>=0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = "src"
module-name = ["remote_read_sql"]
namespace = false

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"
extend-exclude = [".venv"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint]
select = [
    "A",
    "ARG",
    "B",
    "C90",
    "COM",
    "DJ",
    "DTZ",
    "E",
    "F",
    "FA",
    "FURB",
    "I",
    "INT",
    "N",
    "PD",
    "PERF",
    "PL",
    "PLR",
    "PTH",
    "RET",
    "RUF",
    "S",
    "SIM",
    "T20",
    "TRY",
    "UP",
    "W",
    # "FIX",
]
ignore = ["DJ008", "DJ007", "DJ006", "COM812", "PLC0206", "PLR2004", "TRY003"]

[dependency-groups]
docs = [
    "furo>=2025.7.19",
    "sphinx>=8.2.3",
    "sphinx-design>=0.6.1",
    "sphinx-inline-tabs>=2023.4.21",
]
